refactor(flame): extract server routes handling into server-routes.ts
