+++ packages/flame/.docu/node/server.ts
@@ -1,25 +1,20 @@
-import { statSync } from "node:fs";
-import { readFile } from "node:fs/promises";
-import { resolve, join } from "node:path";
 import { watch } from "node:fs";
-import React from "react";
-import { renderToString } from "react-dom/server";
-import { getContentType } from "./utils";
-import { compileMdx } from "./mdx";
-import { DOCS_DIR, DIST_DIR, PAGES_DIR, PROJECT_ROOT, loadDocuConfig } from "./paths";
+import { DOCS_DIR, PAGES_DIR, loadDocuConfig } from "./paths";
 import { loadPlugins } from "./plugin-loader";
 import { BuildPluginBuilder } from "./plugin-builder";
-import type { PageContext } from "./plugin";
-import DocsPage from "../pages/docs/[[...slug]]";
-import NotFoundPage from "../pages/404";
-import IndexPage from "../pages/index";
-import { DocsLayout } from "../components/DocsLayout";
 import { buildClientBundle, computeInlineThemeCss } from "./hydrate";
 import { generateSearchIndex } from "./search-indexer";
 import { logger } from "./logger";
 import { initSentry, captureException } from "./sentry";
-import { SECURITY_HEADERS, generateNonce, isPathSafe, isSlugSafe, htmlResponse } from "./security";
-import { htmlShell as createHtmlShell, hmrScript, errorHtml } from "./html";
+import {
+  serveStatic,
+  handleDocsIndex,
+  handleDocsRoute,
+  handleIndex,
+  handleNotFound,
+  serverErrorResponse,
+  type ServerState,
+} from "./server-routes";
 
 const docuConfig = loadDocuConfig();
 
@@ -52,6 +47,13 @@ if (hasPlugins && builder) {
     await plugin.setup(builder);
   }
 }
+
+const state: ServerState = {
+  docuConfig,
+  assetManifest,
+  inlineThemeCss,
+  builder,
+};
 
 let router: InstanceType<typeof Bun.FileSystemRouter> | null = null;
 try {
@@ -88,216 +90,6 @@ process.on("SIGTERM", () => {
   watcher.close();
   process.exit(0);
 });
-
-function createHtmlResponse(
-  title: string,
-  description: string,
-  body: string,
-  status = 200
-): Response {
-  const nonce = generateNonce();
-  const favicon = docuConfig.meta?.favicon || "/favicon.ico";
-  const html = createHtmlShell({
-    title,
-    description,
-    body,
-    favicon,
-    css: assetManifest.css,
-    js: assetManifest.js,
-    nonce,
-    extraScripts: hmrScript(nonce),
-    themeCss: inlineThemeCss,
-  });
-  return htmlResponse(html, nonce, status, process.env.NODE_ENV !== "production");
-}
-
-async function getDocsForSlug(slug: string) {
-  if (!isSlugSafe(slug, DOCS_DIR)) return null;
-
-  const paths = [
-    join(DOCS_DIR, slug, "index.mdx"),
-    join(DOCS_DIR, `${slug}.mdx`),
-    join(DOCS_DIR, slug, "index.md"),
-    join(DOCS_DIR, `${slug}.md`),
-  ];
-
-  let filePath: string | null = null;
-  let raw: string | null = null;
-  for (const p of paths) {
-    if (!p.startsWith(DOCS_DIR)) continue;
-    try {
-      raw = await readFile(p, "utf-8");
-      filePath = p;
-      break;
-    } catch (err) {
-      if ((err as NodeJS.ErrnoException).code !== "ENOENT") throw err;
-    }
-  }
-  if (!filePath || !raw) return null;
-
-  const relPath = filePath.replace(PROJECT_ROOT + "/", "");
-
-  let content = raw;
-  if (builder) {
-    const transformed = await builder.runOnLoad(relPath, content);
-    if (transformed?.contents) {
-      content = transformed.contents;
-    }
-  }
-
-  const remarkPlugins = builder?.collectRemarkPlugins();
-  const rehypePlugins = builder?.collectRehypePlugins();
-  const result = await compileMdx(content, relPath, undefined, remarkPlugins, rehypePlugins);
-
-  let frontmatter = result.frontmatter as Record<string, unknown>;
-  if (builder) {
-    frontmatter = await builder.runTransformFrontmatterChain(frontmatter, {
-      slug: slug || "/",
-      filePath: relPath,
-      content,
-    });
-  }
-
-  return {
-    content: result.content,
-    compiledSource: result.compiledSource,
-    frontmatter,
-    tocs: result.tocs,
-    filePath: relPath,
-    resolvedContent: content,
-  };
-}
-
-async function renderDocsServerPage(
-  doc: NonNullable<Awaited<ReturnType<typeof getDocsForSlug>>>,
-  slug: string[],
-  pathname: string
-): Promise<Response> {
-  const title = (doc.frontmatter.title as string) || slug.join("/") || "Docs";
-  const description = (doc.frontmatter.description as string) || "";
-
-  const page = React.createElement(
-    DocsLayout,
-    { repoUrl: docuConfig.repo?.url, pathname },
-    React.createElement(DocsPage, {
-      slug,
-      title,
-      description,
-      date: doc.frontmatter.date as string | undefined,
-      content: doc.content,
-      tocs: doc.tocs,
-      filePath: doc.filePath,
-      repoUrl: docuConfig.repo?.url,
-      compiledSource: doc.compiledSource,
-    })
-  );
-
-  const body = renderToString(page);
-
-  if (builder) {
-    const ctx: PageContext = {
-      slug: slug.join("/") || "/",
-      filePath: doc.filePath,
-      frontmatter: doc.frontmatter,
-      content: doc.resolvedContent,
-      config: docuConfig,
-    };
-    const headExtra = builder.collectHead(ctx);
-    const bodyExtra = builder.collectBody(ctx);
-    const nonce = generateNonce();
-    const favicon = docuConfig.meta?.favicon || "/favicon.ico";
-    let html = createHtmlShell({
-      title,
-      description,
-      body,
-      favicon,
-      css: assetManifest.css,
-      js: assetManifest.js,
-      nonce,
-      extraScripts: hmrScript(nonce),
-      themeCss: inlineThemeCss,
-      headExtra,
-      bodyExtra,
-    });
-    html = await builder.runTransformHtmlChain(html, ctx);
-    return htmlResponse(html, nonce, 200, process.env.NODE_ENV !== "production");
-  }
-
-  return createHtmlResponse(title, description, body);
-}
-
-async function handleDocsIndex(): Promise<Response> {
-  const doc = await getDocsForSlug("");
-  if (!doc) return renderPage(NotFoundPage, "404 - Not Found", "", 404);
-  return renderDocsServerPage(doc, [], "/docs");
-}
-
-async function handleDocsRoute(slug: string[]): Promise<Response> {
-  const path = slug.join("/");
-  const doc = await getDocsForSlug(path);
-  if (!doc) return renderPage(NotFoundPage, "404 - Not Found", "", 404);
-  return renderDocsServerPage(doc, slug, `/docs/${path}`);
-}
-
-function renderPage(
-  Component: React.ComponentType<Record<string, unknown>>,
-  title: string,
-  description: string,
-  status: number,
-  props: Record<string, unknown> = {}
-): Response {
-  const page = React.createElement(
-    DocsLayout,
-    { repoUrl: docuConfig.repo?.url, pathname: "/docs" },
-    React.createElement(Component, props)
-  );
-  const body = renderToString(page);
-  return createHtmlResponse(title, description, body, status);
-}
-
-function handleIndex(): Response {
-  const page = React.createElement(IndexPage);
-  const body = renderToString(page);
-  return createHtmlResponse(
-    docuConfig.meta?.title || "DocuBook",
-    docuConfig.meta?.description || "",
-    body
-  );
-}
-
-function serveStatic(pathname: string): Response | null {
-  if (!isPathSafe(pathname, DIST_DIR)) return null;
-  const decoded = decodeURIComponent(pathname);
-  const assetPath = resolve(DIST_DIR, decoded.slice(1));
-  try {
-    const s = statSync(assetPath);
-    if (s.isFile()) {
-      return new Response(Bun.file(assetPath), {
-        headers: { "Content-Type": getContentType(pathname) },
-      });
-    }
-  } catch (err) {
-    if ((err as NodeJS.ErrnoException).code !== "ENOENT") throw err;
-  }
-
-  if (decoded.startsWith("/docs/assets/")) {
-    const docsAsset = resolve(DOCS_DIR, "assets", decoded.replace("/docs/assets/", ""));
-    const docsAssetsDir = resolve(DOCS_DIR, "assets");
-    const docsAssetsDirSlash = docsAssetsDir.endsWith("/") ? docsAssetsDir : docsAssetsDir + "/";
-    if (docsAsset !== docsAssetsDir && !docsAsset.startsWith(docsAssetsDirSlash)) return null;
-    try {
-      const s = statSync(docsAsset);
-      if (s.isFile()) {
-        return new Response(Bun.file(docsAsset), {
-          headers: { "Content-Type": getContentType(pathname) },
-        });
-      }
-    } catch (err) {
-      if ((err as NodeJS.ErrnoException).code !== "ENOENT") throw err;
-    }
-  }
-  return null;
-}
 
 const server = Bun.serve({
   port: PORT,
@@ -361,19 +153,19 @@ const server = Bun.serve({
           const slug = slugParam ? slugParam.split("/") : [];
 
           if (slug.length === 0) {
-            response = await handleDocsIndex();
+            response = await handleDocsIndex(state);
           } else {
-            response = await handleDocsRoute(slug);
+            response = await handleDocsRoute(slug, state);
           }
         } else if (routeName === "/404") {
-          response = renderPage(NotFoundPage, "404 - Not Found", "", 404);
+          response = handleNotFound(state);
         } else if (routeName === "/") {
-          response = handleIndex();
+          response = handleIndex(state);
         } else {
-          response = renderPage(NotFoundPage, "404 - Not Found", "", 404);
+          response = handleNotFound(state);
         }
       } else {
-        response = renderPage(NotFoundPage, "404 - Not Found", "", 404);
+        response = handleNotFound(state);
       }
 
       logger.request(
@@ -386,32 +178,14 @@ const server = Bun.serve({
     } catch (err) {
       captureException(err, { method: req.method, pathname });
       logger.request(req.method, pathname, 500, Math.round(performance.now() - startTime));
-      const msg = err instanceof Error ? err.message : String(err);
-      const st = err instanceof Error ? err.stack : undefined;
-      return new Response(errorHtml(msg, st), {
-        status: 500,
-        headers: {
-          "Content-Type": "text/html",
-          ...SECURITY_HEADERS,
-          "Content-Security-Policy": "default-src 'none'; style-src 'unsafe-inline'",
-        },
-      });
+      return serverErrorResponse(err);
     }
   },
 
   error(error) {
     console.error(error);
     captureException(error);
-    const msg = error?.message ?? "Unknown error";
-    const st = error?.stack;
-    return new Response(errorHtml(msg, st), {
-      status: 500,
-      headers: {
-        "Content-Type": "text/html",
-        ...SECURITY_HEADERS,
-        "Content-Security-Policy": "default-src 'none'; style-src 'unsafe-inline'",
-      },
-    });
+    return serverErrorResponse(error);
   },
 });
